+++ src/utils/generate-icon.js
